+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Support for multiline imports ([#60](https://github.com/buehler/typescript-hero/issues/60))
 - Added setting for multiline threshold
 - Configurable new import location (at top of the file or at the cursor position) ([#41](https://github.com/buehler/typescript-hero/issues/41))
+- Asks for alias if a specifier is already present ([#44](https://github.com/buehler/typescript-hero/issues/44))
 
 #### Fixed
 - Autocomplete does not suggest items that are already imported ([#64](https://github.com/buehler/typescript-hero/issues/64))